apfsck: check refcnts even for extents with none
For each physical extent, e_total_refcnt is currently calculated when we
encounter a reference to it. When no references are found, it's left as
zero, so the reference check in free_extent() passes without actually
confirming anything. Instead calculate e_total_refcnt right before the
check, inside free_extent(). This approach gives extents with zero total
references the same treatment as the others.

The only problem is that we are redoing queries that had already been
performed before, so there will be a performance cost. I don't care much
about speed here, so I haven't bothered to check how big it is. Probably
not much.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>