apfsck: allow all inodes to share id with dstreams
I used to think that inodes could only possibly share an id with their
own dstream. This was a mistake: even after a cloned inode is modified
(changing its dstream id), its clone may still point to the original
dstream, which continues to share the original inode's id. Adjust the
cnid reuse checks to allow this.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>