apfsck: fix checks for orphan inodes
I was under the impression that orphan files followed the same rules as
all other files, just inside the private directory. This turned out to
be wrong: they actually have a link count of zero, and they preserve the
parent name and id for their original location. Adjust the checks.

The kernel driver will also need to be updated eventually, if we want it
to behave just as the official implementation. According to the
reference, this is not strictly required, but it might be better that
way.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>