apfsck: allow missing extents for orphan inodes
I've encountered an image that has an orphan inode with a large reported
size, but without any actual extents. My impression is that this is the
way large files are deleted after they can no longer be recovered, so it
should be possible to find inodes that are missing only some extents.

Allow any amount of missing extents, at least for now, but leave a
warning in place so that I notice quickly if I ever find a partially
deleted orphan.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>