apfsck: fix minor bug in attach_extent_to_dstream
The loop in attach_extent_to_dstream() is supposed to continue after an
extent is found to already be counted for this owner. I'm pretty sure
this makes no difference (if one extent is already considered then they
all are), but I get very confused with this stuff so fix it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>