apfsck: fix num_snapshots check inside snapshots
The apfs_num_snapshots field of a snapshot's volume superblock is not
always set to zero. The value at the time of the snapshot creation is
kept permanently, so it should actually be equal to the number of other
snapshots encountered so far - except, of course, if one of the older
snapshots was deleted. Fix this.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>