apfsck: allow future xids in old extentref trees
When a snapshot is deleted, its physical extent records get moved to the
extentref tree for the next one (or to the current tree if there are no
other snapshots). This is done under the current transaction, so nodes
in that tree will have xids that appear to be from the future. Allow it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>