apfsck: ignore extended snap meta for current xid
I'm not sure why the current xid gets extended snapshot metadata. My
guess is that it just gets copied along with everything else when a
snapshot is created, which explains why it usually matches the latest
snapshot's, and why it is not updated when that snapshot gets deleted.

So, it's irrelevant and impossible to check in any meaningful way. Leave
only the most superficial checks in place, and remove the rest.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>