apfsck: check crypto state records
Implement some basic checks for crypto state records, which were the
only remaining unsupported record kind. For now we focus only on file
encryption and ignore metadata encryption, which I'm yet to encounter.

While testing this patch I found that some cloned dstreams don't have a
crypto id of their own: they set it to -1 and apparently use the one for
the original file. I'm not sure how any of this works yet.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>