apfsck: allow a single leaked block in the omap
For reasons I ignore, some ios image volumes seem to have an orphaned
extended snap metadata block. It's listed in the volume's object map,
but it's not referenced in any of the snapshots as far as I can see.
Allow this for now, maybe I'll make sense of it at some point.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>