apfsck: support the overprovisioning file
Images from iOS devices have something called an "overprovisioning
file", apparently for the sake of their NAND storage. For some reason,
it seems that this file can have a wrapped crypto state record with no
actual key, so we mistake it with a metadata crypto record and report it
as unknown.

So, don't run the nonempty crypto key check for these files. Also add
some superficial checks for the filename and link count, though I don't
know for sure if they are correct or required.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>