apfsck: update the container superblock layout
It seems that I missed a change in the layout of the container
superblock, with two new fields added to the end of it. One of those
fields, the wrapped media key, refers to an allocated block, so take it
into consideration for the spaceman checks.

This is finally enough to get those checks to pass for at least one of
my testing images.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>