apfsck: support newer xfield types
I've encountered a directory that has the INO_EXT_TYPE_PURGEABLE_FLAGS
xfield. The existence of this xfield type is documented, but the value
is claimed to be reserved, and no details are given. The reference seems
to imply that only purgeable files should use this, but the directory I
saw was not purgeable itself (though it did include mostly purgeable
files).

Anyway, for now just add superficial size checks for this xfield type,
and also for INO_EXT_TYPE_ORIG_SYNC_ROOT_ID. Both had been missing from
the original reference and were added later on.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>