apfsck: add checks for another purgeable drec flag
I have encountered a second dentry flag that purgeable files may have. I
don't understand what makes it different from the original, but then
again I don't understand purgeable files, so it's not surprising. I do
notice that the reported filetype value changes depending on which flag
is in use, so add checks for that.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>