apfsck: fix bug in current allocation zone checks
We currently check azb->saz_zone_start and azb->saz_zone_end before
actually setting azb in the loop, so it keeps the stale value from the
previous loop and triggers a bogus report. Fix it, and follow my current
policy of initializing all pointers to NULL, which would have prevented
this in the first place.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>