apfsck: allow reserved type to be set on volume
I have encountered an image that has a value in the reserved_type field
of the volume superblock, even though the matching reserved_oid is set
to zero. Object types are always reported but so far I haven't seen them
change, so I assume this one will always have the same value as well.
Add a check for it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>