apfsck: check that cloneinfo fields are trivial
After checking the reference with more care, the cloneinfo epoch and
cloneinfo xid are actually explained in detail. They are a mitigation
for a bug in older versions of the official apfs implementation, so that
filesystems that got corrupted by them can still be used safely in newer
software.

This doesn't matter much to me because I don't think I'll be checking
the potentially corrupted INODE_WAS_EVER_CLONED in the driver, I'll just
check the extent/dstream refcounts. I might change my mind when I get to
work on it, though.

For now just demand that these fields have trivial values, that is, that
the volumes are either older than the mitigation or newer than it. We
completely ban volumes that were mounted first with good implementations
and later with buggy ones, which is hopefully rare. We report the
reverse, which is probably seen in quite a few early disks, as unknown.
I may add better checks if I ever see one of these images, but like I
said before, I don't believe it will matter for the driver.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>