apfsck: parse the container keybag
One of the apfs images I use for testing has a single allocated block
that I have struggled to account for. Now I finally understand where it
comes from: it's the block for a volume keybag, and those are listed
encrypted inside the contaner's keybag. So the only way to finally get
compare_container_bitmaps() to pass is to actually decrypt and parse the
keybag.

The official reference claims that RFC 3394 is in use here, but that's
not true: it actually uses AES under the XTS mode of operation. I got
that info from the apfs-fuse <https://github.com/sgan81/apfs-fuse>
source code. This patch includes an implementation of AES which I ripped
from wpa_supplicant, so it's under the (3-clause) BSD license. I've also
included their implementation of RFC3394, even though it's not currently
needed.

As for the XTS implementation, I intended to write an original one in
case I ever need to avoid the GPL license from apfs-fuse. This did not
work out, because I had that version too fresh in my mind, so I ended up
mostly doing a C port of that code.

Anyway, once the container keybag is decrypted, run superficial checks
on the few tags I've encountered so far and mark the volume keybag
blocks as used. Do not actually parse and check the volume keybag for
now.

Strangely enough, I've encountered at least one container keybag that
doesn't appear to be encrypted at all. Put provisions in place for that
case as well.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>