apfsck: tweak allocated block count for snapshots
I've had some trouble getting allocated block counts to add up in the
presence of snapshots. On images produced by a recent iOS release, the
check would pass if there was a single snapshot, but each new snapshot
added would make the calculated count drift by one.

I think I've figured this out, and it's a combination of two problems.
First, there is the leaked block I mentioned in 12731e61 ("apfsck: allow
a single leaked block in the omap"). Even if it doesn't seem to be
reachable, it still needs to be counted as part of its volume. Ignoring
this block made the check appear to pass for the one-snapshot case.

The second problem are the volume superblocks. I've always known that
they were not counted, but the previous confusion made me think that
they behaved differently inside snapshots. They don't though, and they
always need to be ignored.

Fix both issues.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>