apfsck: allow the regular file count not to match
I've always had an image where the reported regular file count seems to
be off by one. After I ran into a similar issue with another image, I
fixed it with commit 8b1575bd ("apfsck: ignore orphans in the volume
file counts") and I thought that was it. But no, the original image is
still failing and I have no idea why.

Make this failure into a "weird" report again, like it was before that
commit.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>