apfsck: explain the leaked SNAP_META_EXT block
In the past I've encountered extended snapshot metadata blocks listed in
the omap, but unreachable from any of the snapshot volumes. I added some
checks to allow this in commit 12731e61 ("apfsck: allow a single leaked
block in the omap"), but I had no idea why it was happening.

While I was implementing this stuff in the driver, I finally figured it
out. The extended snapshot metadata was a late addition to apfs, and the
official reference allows old implementations to simply ignore it when
they find it. So, when a snapshot is created by a recent implementation
and then deleted by an old one, this block would leak. This is fine and
not a form of corruption.

Now that I understand this, update the comments on the matter. Also
allow any number of leaked blocks of this type: so far I've only ever
seen one, but there's no limit here.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>