apfsck: fix inline xattr compression for nonsealed
For compressed files in sealed volumes, we keep a list of hashes of the
contents inside the xattr's dstream struct. Since some of those xattrs
are inline, they use a fake dstream struct with a d_inline field set.
When I implemented this, I forgot to cover the case of nonsealed
volumes, which won't go through any of this, so now they are all broken.
Fix it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>