apfsck: don't leak maps in read_whole_dstream()
I don't care much about leaks in the fsck, at least for now. It's just
intended as a hacky tool for testing the driver, and some of the leaks
would be tricky to fix without big rewrites. This one though, is very
dumb, and it's so brutal that we are actually hitting vm.max_map_count
for some images. Fix it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>