Fix double check of objects during dstream read
Some images with compressed files are failing with "oid used twice for
same snapshot". The reason is that read_whole_dstream() will inevitably
go through some objects a second time during the queries. I already have
the 'ongoing_query' global to deal with these situations, but I forgot
to use it when adding the snapshot checks. Fix it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>