apfsck: fix checks of index size for fixed kv size
After the driver creates a snapshot, the official fsck complains about
the size of a node's table:

  ** Checking the snapshot metadata.
  error: btn: invalid btn_table_space (0, 64), given btn_flags (0x7)

This fsck doesn't seem to notice. There are two bugs that conspire to
hide this kind of corruption. One is that nodes with fixed key/value
sizes are supposed to have an index of the exact size to map the whole
block, but we currently treat this number only as a minimum. The other
is that I forgot to add support for OMAP_SNAPSHOT and FEXT tree types
to node_min_table_size(), so those trees are treated as variable-length.
Fix both problems.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>