apfsck: allow bigger index size for free queues
While free queues have fixed key/value sizes, they are different than
other such trees in that they also use "ghost records", which don't have
a value at all. The point of this is to save space, so it makes no sense
if we force the nodes to have a fixed record count. The driver always
did this correctly, but this fsck has started complaining about it with
the previous patch. Fix it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>