apfsck: attempt to handle extent refcnts correctly
Now that I've actually implemented snapshots in the driver, I can see
that the way extent reference counting is handled in the fsck makes no
sense at all. Update extents mean that the number of references can be
different for each block of a physical extent. So now, the extents
listed in the table will get split when it becomes necessary, and each
fraction will have its own refcnt.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>