apfsck: check that freed blocks <= alloced blocks
After making some changes to one of my images with the official driver,
the official fsck complains that the reported number of freed blocks in
a volume became higher than the reported number of alloced blocks. This
is not considered an error, just a warning, but still add the same check
here. I think it's the only check that makes sense for these two fields.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>