apfsck: check consistency of inode clone flags
On a clone, the resulting file gets the APFS_INODE_WAS_CLONED flag set.
The APFS_INODE_WAS_EVER_CLONED sounds like it should only get set on the
original file, but it actually gets set on both. So, every file with
APFS_INODE_WAS_CLONED should also have APFS_INODE_WAS_EVER_CLONED. I'm
not entirely sure about this, but add a check for it and wait to see if
something breaks.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>