apfsck: fix report for blocks with two NEW records
In my testing of the driver, I ran into a situation where a single block
had two NEW extent records. The fsck reported this as a bug in its own
code, but that's not correct of course: it was a bug in the driver.
Report it correctly from now on.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>