Don't hardcode the sm offset of the cib address
We store the cib address inside the spaceman, right after
ip_bm_free_next. I don't think this is mandatory, but I have never tried
to put it anywhere else. Right now we have a problem, because the size
of ip_bm_free_next depends on the number of ib bitmap blocks, so for big
disks we overwrite the cib address when ip_bm_free_next is generated. So
stop hardcoding the cib address offset and calculate it from the ip
size.

By the way, the apfsck doesn't report this problem correctly. Instead,
it throws SIGBUS because we are mmapping a huge block number for the
cib. I don't think this is very important right now, but it could be
improved.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>