apfsck: fix check for multiblock ip_bitmap_offset
The spaceman field whose offset is given by sm_ip_bitmap_offset is
actually not a single 64-bit number, but an array of 16-bit numbers,
one for each ip bitmap block. This is all very similar to the previous
problem with xids, fixed in e3926bbf5632 ("apfsck: check xids for
multiblock ip bitmaps").

In this case, the array tells us which of the bitmap blocks from the
nonfree region are actually in use, and in what order they should be
considered. So handle the ip bitmap properly, reading the 16-bit offsets
one by one.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>