mkapfs: set multiblock ip_bitmap_offset correctly
As mentioned before in 50eb66c18473 ("apfsck: fix check for multiblock
ip_bitmap_offset"), the field pointed to by sm_ip_bitmap_offset is not
actually a single 64-bit number, but an array of 16-bit numbers, one for
each bitmap block in use for the internal pool. In our case, we are just
using consecutive blocks (the first few in the ring), so set the array
to consecutive numbers starting from 0.

Also allow the array to take more than 8 bytes, since I imagine that
will be needed for bigger images (I haven't tested this yet). Keep the
following field aligned to 8 bytes though.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>