apfsck: don't check tails of multiblock ip bitmaps
I just attempted to create an image with a size of 6T, so that it would
need 5 bitmaps to map the internal pool. I haven't checked it with the
official fsck yet, but mine chokes on it:

  Internal pool: non-zeroed bitmap.

This comes from a silly check I have in place, in a function called
check_ip_bitmap_blocks(). This goes through all the stale ip bitmaps in
the ring, and confirms that none of them has bits set that map to blocks
after the end of the internal pool. I guess this was intended as a
little sanity check, to make sure I understood what the old bitmaps were
doing.

But this doesn't make any sense for multiblock bitmaps, because bitmap
blocks that don't map the tail blocks of the ip will be used in full.
When they go stale, there is no way to tell them apart from the tails.
Just refuse to run this check for the multiblock case.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>