apfsck: implement checks for cib address blocks
Large containers don't list CIBs directly in the spaceman, they instead
list something called a cib address block (CAB) which itself contains a
list of CIBs. Add checks for this case. It's straightforward because the
logic is more or less the same as for CIBs.

This has all been tested on a single fresh 10Tb image created with the
official newfs_apfs. Much more testing will be necessary, but I'll
probably wait until CABs are implemented in the driver.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>