apfsck: support multiblock ephemeral objects
For small filesystems, the addresses of the CIBs are kept as an array in
the spaceman block. For huge filesystems, the spaceman instead lists the
addresses of CABs, and those in turn list the CIBs. One might expect the
boundary between the two to be before the point where the CIBs overrun
the spaceman block, but that's not correct.

I used the official newfs_apfs to create a filesystem with 8574096900096
bytes (~7.8T), which is the maximum size that would fit in a single CAB.
As it turns out, this filesystem still keeps the CIB list inside the
spaceman, but the length of the spaceman object is now two blocks
instead of one.

I had never encountered a multiblock ephemeral object before, but I
always suspected they were possible because the object size is reported
in checkpoint mappings. This is a bit annoying because pretty much all
objects are a single block in size, and now I need to pass the object
size around everywhere just for this one crazy spaceman case. There is
no way around it though, so just do it.

By the way, filesystems bigger that 8574096900096 bytes do use CABs as
expected. That's where the limit is: containers that would fit in a
single CAB don't use it at all. This had been my guess in the previous
patch, I just failed to realize that the CIB list would overrun the
spaceman block.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>