mkapfs: support cib address blocks
As explained in commit daa19ddc6f75 ("apfsck: implement checks for cib
address blocks"), containers with more than 8574096900096 bytes list
cib address blocks (CABs) in the spaceman instead of CIBs. Support this
in the mkfs.

With this change, I believe we've finally removed all userland hurdles
to support large containers, and I can move on to implement them in the
driver. I did leave an upper size limit of around 3.6 PB, so that all
CAB addresses would continue to fit in a single spaceman block. This
would be easy to fix (and the official mkfs does it fine), but I don't
think there is any point.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>