mkapfs: add tests for different container sizes
The space manager of an apfs container changes a lot depending on the
size of the filesystem. I fear that I will forget which size ranges are
of special interest for testing so, as a way to document this, put a
simple test script in place here. Just check that the fsck and mkfs are
consistent for all interesting size ranges I support so far.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>