mkapfs: report release info in apfs_formatted_by
The filesystem superblock has an apfs_formatted_by field to report which
mkfs implementation was responsible for creating the container in the
first place. Improve the version information provided, following the
recent driver changes from commit e0b1eeb29d2a ("Include release info in
apfs_modified_by field").

One small difference to note with the driver case is that we need to use
the "--always" flag with "git describe", or else it fails. The reason is
that I've never made a tagged release here. I should get to that soon.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>