Report release info with "-v" flag
Make all tools report the current git commit id (or tag, if any) when
called with the "-v" flag. This will be useful to confirm I'm not
running an old version by mistake.

The makefile changes here are a bit awkward. In general this repo's
structure isn't great. I should look into that at some point.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>