mkapfs: correct reported upper size limit
I don't know what went wrong with my previous mkapfs tests that made me
think we could support filesystems up to 3.6 PiB. In reality, the limit
is around 113 TiB. The code is still correct because it doesn't use any
hardcoded value for the size check, but fix the comments and the error
message.

By the way, the limit is much lower than I would have expected because
the metadata for the ip bitmaps takes up much more space than the CAB
addresses, and I had forgotten to take those into account. Mention it in
the comments.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>