Ignore orphans for directory statistics
If both an inode and its parent directory get deleted, and then the
parent gets evicted from the orphan list before the child, the child
will continue to report the same parent id referring to the now
nonexistent inode. New changes to my driver are making this possible,
and as a result collect_dirstats() creates an empty inode struct for the
old parent, and later reports that it has "no name for primary link".

Anyway, the problem here is that collect_dirstats() is considering
orphan files at all. The parent id they report is always stale, and the
stats for their old parent should have already been updated with the
deletion, so just ignore them entirely.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>