mkapfs: support containers of less than 128M
I remember reading in the official reference that containers under 128M
were a special case. At the time I assumed this was because the whole
disk could get mapped in a single chunk, and so I expected some
complications. But I've been looking at a tiny image for the first time
today, and there's actually no real work that needs to be done. Just set
the proper value in the ephemeral info field.

With this patch, the lower size limit is pushed to 83M. I can't make it
smaller yet because of the hardcoded block numbers, but I'll look into
that next.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>