mkapfs: don't hardcode checkpoint desc/data sizes
The sizes of the checkpoint descriptor and data areas have always been
hardcoded here. This is probably a bad idea in general, but currently it
makes it impossible to support tiny filesystems. So calculate the sizes
from the container block count instead.

I tried to test the official newfs_apfs to figure out their formula,
which has probably been somewhat optimized. It's a bit messy though, and
I doubt the exact values matter, so I ended up with a rough
approximation.

With this change the minimum supported container size drops to 512K,
which is as far as I intend to go. Update the test script as well.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>