mkapfs: fix size of spaceman mmap
When we map the spaceman blocks into memory with get_zeroed_blocks(), we
are supposed to pass the spaceman size in blocks as a parameter. Instead
I mistakenly passed the size in bytes, as returned by spaceman_size().
This has not been a problem so far, but it will trigger sigbus for tiny
filesystems. Fix it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>