mkapfs: fix checkpoint data blocks formula for 3G
For containers of 3G, mkfs is currently failing with sigbus. This is a
result of the previous patch, which uses an unnecessarily convoluted
formula for the number of data blocks in filesystems between 1G and 4G,
as part of a senseless (and failed) attempt to get the exact same values
as the offcial newfs_apfs. The problem is that a term of the formula can
go negative for some sizes, and C integer promotion rules turn this into
huge values. I should really simplify the formula, but I'm really tired
of this so for now just cast the dangerous term to a big enough signed
type.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>