mkapfs: remove comment about oldest xid in fq
There are two very old comments in the functions that create the free
queues. It seems that, at the time, I had doubts about the proper value
for sfq_oldest_xid on an empty free queue. Using zero has caused no
trouble at all so it's time to clean this up.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>