apfsck: ignore WAS_EVER_CLONED before the epoch
So far I have mostly ignored this "cloneinfo id epoch" in the fsck, but
I notice now some issues with one of my older images. I guess it was
created before the bugfix, so some inodes have the WAS_CLONED flag, but
not WAS_EVER_CLONED. Since this is technically allowed stop reporting it
as corruption, but leave a warning message in place anyway.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>