apfsck: add cli option to specify a tier 2 device
As a first step towards supporting fusion drives, add a command line
option to specify the tier 2 device. We now have two global fd's in use,
so give them names that are less confusing. As a simple sanity check,
fix the block count check, which is the first to fail if the fsck gets
run against the main device alone.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>