mkapfs: run tests on fusion drives as well
Expand the existing test script to run tests on fusion drives as well,
going through all possible combinations of the size array. Of course
these specific sizes are not as relevant for fusion drives, because they
were picked as the critical points where the layout changes for regular
drives, but hopefully it will be enough to cover everything.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>