Avoid sigbus during fusion drive tests
After commit f523ce46 ("mkapfs: run tests on fusion drives as well"), I
noticed that a few fusion drive tests are crashing with sigbus, but I
got sidetracked by another issue and I only got to it now. The problem
is, of course, that with a large tier 2 disk the metadata may not fit in
the main device. This is not a bug, but it's a bit ugly to fail with
sigbus, so put some coarse space checks in place and fail cleanly when
the size difference is too crazy.

Also update the fusion drive tests so that they start checking for clean
failure in the odd cases.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>