apfsck: fix out-of-bounds write from last patch
The used_bitmaps array was allocated with one slot for each used ip
bitmap block, but it should also have a slot for the free tail. This
isn´t serious because I don't expect anyone to use this fsck other than
me (lots of similar issues probably exist), but it's causing the mkfs
tests to fail, so fix it.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>