apfsck: check the block zero magic before all else
When I mistakenly run the fsck on a whole disk, instead of running it on
the apfs partition, I get a confusing "block size is too small" error
message. To avoid this noise, run no checks at all until we have
confirmed that we are working with an apfs filesystem. If not, explain
this clearly to the user.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>